Changed "evt" to "e" for style preference
In the rest of Flight's codebase we always either use the whole word "event" or the shorthand "e" for the event being emitted. I went with "e" since you were using a shorthand. This is not to say there's any actual problem with calling it "evt" if you worked alone or with a team that preferred that name for events.